pyproject.toml
[project]
name = "familien-dashboard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "postgres>=4.0",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "gunicorn>=23.0.0",
    "requests>=2.32.0",
]
